todo/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
